gpu: drm: adv7511-legacy: drop driver since it's no longer used
This driver seems to have been around for a while, but for some time we are
using the upstream version.

This driver isn't even compiled, since it was added to the
'compile_exceptions' lists.

So, remove the driver. No need to keep it around.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit dc367ff92c22b44d81c830543b4f0e618af01a1b)