media/axi-hdmi-rx: Honor pclk-sample from devicetree
The driver defaults to sample on falling edge, but did not offer an interface
to change this. Implement the "pclk-sample" as specified in the devicetree
bindings (video-interfaces.txt) to work for this device.

This solves the artefacts caused by timing issues we've seen on all our boards.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>
(cherry picked from commit f7dfc9fe58ddbd58603488f557075b2bc513ca40)